refactor(e2e): refactor onboarding flow tests
